Routing Service initial implementation

+++ utilities/src/test/java/com/orbyte/utils/UniqueGeneratorTest.java
@@ -2,6 +2,7 @@ package com.orbyte.utils;
 
 import com.orbyte.dto.PaymentKeyDTO;
 
+import java.math.BigInteger;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.time.LocalDateTime;
@@ -15,7 +16,7 @@ public class UniqueGeneratorTest {
 
     @Test
     public void testGenerateUniquePaymentKeyPositive() throws NoSuchAlgorithmException, InvalidKeyException {
-        PaymentKeyDTO paymentKey = PaymentKeyDTO.builder().secret("testsecret").amount(4000).currency("USD").transactionDate(LocalDateTime.now()).build();
+        PaymentKeyDTO paymentKey = PaymentKeyDTO.builder().secret("testsecret").amount(BigInteger.valueOf(4000)).currency("USD").transactionDate(LocalDateTime.now()).build();
 
         Assertions.assertDoesNotThrow(()->uniqueGenerator.generateUniquePaymentKey(paymentKey));
